move api calls from details to registeration
moved all the api calls from Player Details to Account Registration so that when the user makes an account the api loads there. and when details is accessed, it uses the DB instead of api calls.
However it is not being saved in the database yet. That will be the next step

+++ EScouting/EScouting/Models/Match.cs
@@ -40,10 +40,10 @@ namespace EScouting.Models
         public string summonerName { get; set; }
         public string matchHistoryUri { get; set; }
         public string platformId { get; set; }
-        public int currentAccountId { get; set; }
+        public long currentAccountId { get; set; }
         public int profileIcon { get; set; }
         public int summonerId { get; set; }
-        public int accountId { get; set; }
+        public long accountId { get; set; }
     }
 
     public class Team
