+++ EScouting/EScouting/Models/Summoner.cs
@@ -11,7 +11,7 @@ namespace EScouting.Models
         public int profileIconId { get; set; }
         public string name { get; set; }
         public int summonerLevel { get; set; }
-        public int accountId { get; set; }
+        public long accountId { get; set; }
         public int id { get; set; }
         public long revisionDate { get; set; }
     }
